Feature: Scenes Added basic scenes managment
