+++ camera/camera.go
@@ -1,4 +1,4 @@
-package main
+package camera
 
 import "math"
