+++ main.go
@@ -1,44 +1,10 @@
 package main
 
 import (
-	"image"
 	"log"
-	"projectEVA/constants"
-	"projectEVA/entities"
 
 	"github.com/hajimehoshi/ebiten/v2"
 )
-
-func CheckCollisionHorizontal(sprite *entities.Sprite, colliders []image.Rectangle) {
-	for _, collider := range colliders {
-		if collider.Overlaps(image.Rect(
-			int(sprite.X),
-			int(sprite.Y),
-			int(sprite.X)+constants.Tilesize,
-			int(sprite.Y)+constants.Tilesize)) {
-			if sprite.Dx > 0.0 {
-				sprite.X = float64(collider.Min.X) - constants.Tilesize
-			} else if sprite.Dx < 0.0 {
-				sprite.X = float64(collider.Max.X)
-			}
-		}
-	}
-}
-func CheckCollisionVertical(sprite *entities.Sprite, colliders []image.Rectangle) {
-	for _, collider := range colliders {
-		if collider.Overlaps(image.Rect(
-			int(sprite.X),
-			int(sprite.Y),
-			int(sprite.X)+constants.Tilesize,
-			int(sprite.Y)+constants.Tilesize)) {
-			if sprite.Dy > 0.0 {
-				sprite.Y = float64(collider.Min.Y) - constants.Tilesize
-			} else if sprite.Dy < 0.0 {
-				sprite.Y = float64(collider.Max.Y)
-			}
-		}
-	}
-}
 
 func main() {
 	ebiten.SetWindowSize(1920, 1080)
